+++ CHANGELOG.md
@@ -7,6 +7,8 @@ v0.6.4 (--- -- ----)
 
 ## Fixed
  * Check on the argument count on callbacks on adding them.
+ * Now the on-function returns the callback it was given.
+
 
 v0.6.3 (May 5 2022)
 ===============================================================================